requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow==2.15.0
Pillow==8.4.0
streamlit 